requirements/tests.txt
pytest==5.4.3
pytest-cov==2.9.0
pytest-watch==4.2.0
pytest-mock==3.1.0
pytest-asyncio==0.12.0
pytest-click==0.3
pytest-randomly==3.4.0
pytest-aiohttp==0.3.0
mock==4.0.2

asyncpg==0.20.1

pytest-mypy-plugins==1.3.0
mypy==0.770
